% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_definitions}
\alias{annotate_definitions}
\title{annotate all input defentions in code blocks WIP}
\usage{
annotate_definitions()
}
\description{
identify and annotate all defenitions within code block. WIP
}
