% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_side_effects}
\alias{annotate_side_effects}
\title{annotate all side effects for each code block}
\usage{
annotate_side_effects(blocks, reactives)
}
\arguments{
\item{blocks}{\code{data.frame} all code blocks in server.R code}

\item{reactives}{\code{vector} of all reactives, both reactiveVal() and values$reactiveValues}
}
\description{
identify and annotate all side effects triggered in code block.
  Take both reactiveVal() and values$reactiveValues into account
}
