% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_type}
\alias{annotate_type}
\title{annotate all inputs/reactives for each code block}
\usage{
annotate_type(blocks, inpts, type = c("inputs", "reactives"))
}
\arguments{
\item{blocks}{\code{data.frame} all code blocks in server.R code}

\item{inpts}{\code{vector} of all inputs/reactives}

\item{type}{\code{string} which type to annotate: inputs or reactives}
}
\description{
identify and annotate all inputs/reactives referenced in code block.
  Take isolated inputs/reactives into account
}
