% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{compute_dependency_network}
\alias{compute_dependency_network}
\title{compute source, target dependency network}
\usage{
compute_dependency_network(app, type = c("inputs", "reactives"),
  relation = c("dependency", "isolate"))
}
\arguments{
\item{app}{\code{data.frame} all annotated code blocks in server.R code}

\item{type}{\code{string} which type to annotate: inputs or reactives}

\item{relation}{\code{string} which type to relation: dependency or isolate}
}
\description{
create a source, target data.frame listing all dependencies and their direction
}
