% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{compute_network}
\alias{compute_network}
\title{compute source, target network of (isolated) dependencies and side effects}
\usage{
compute_network(app)
}
\arguments{
\item{app}{\code{data.frame} all annotated code blocks in server.R code}
}
\description{
create a source, target data.frame listing all dependencies and side effects and their direction
}
