% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{compute_side_effect_network}
\alias{compute_side_effect_network}
\title{compute source, target side effects network}
\usage{
compute_side_effect_network(app)
}
\arguments{
\item{app}{\code{data.frame} all annotated code blocks in server.R code}
}
\description{
create a source, target data.frame listing all side effects and their direction
}
