% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid.R
\name{convert_to_mermaid_connection}
\alias{convert_to_mermaid_connection}
\title{generate mermaid code for single dependency}
\usage{
convert_to_mermaid_connection(source, target, id, relation)
}
\arguments{
\item{source}{id of source as assigned by \code{\link{parse_code_blocks}}}

\item{target}{id of target as assigned by \code{\link{parse_code_blocks}}}

\item{relation}{\code{string} type of relation ("isolate" or "dependency")}
}
\description{
generate mermaid code for single dependency
}
