% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid.R
\name{convert_to_mermaid_node}
\alias{convert_to_mermaid_node}
\title{generate mermaid code for single reactive block}
\usage{
convert_to_mermaid_node(id, name, type, color)
}
\arguments{
\item{id}{\code{string} id of block as assigned by \code{\link{parse_code_blocks}}}

\item{name}{\code{string} assigned name of block (\code{(output$)<name> <- reactive/render*})}

\item{type}{\code{string} type of block (reactive, observeEvent, ...)}

\item{color}{\code{string} color block should get in mermaid graph}
}
\description{
generate mermaid code for single reactive block
}
