% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{delete_isolates}
\alias{delete_isolates}
\title{delete isolate block from within code block expressions}
\usage{
delete_isolates(code, type, regex = regex_pattern()$isolate)
}
\arguments{
\item{code}{\code{string} expression of code block}

\item{type}{\code{string} type of code block (observe, reactive, observeEvent, ...)}

\item{regex}{\code{string} regex pattern to detect isolation blocks}
}
\description{
detect and delete isolate blocks from code block expression
}
