% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{determine_invalidators}
\alias{determine_invalidators}
\title{determine reactive invalidators}
\usage{
determine_invalidators(type, expr, opts)
}
\arguments{
\item{type}{\code{string} type of code block (reactive, observeEvent, ...)}

\item{expr}{\code{vector} inputs/reactives referenced in main expression}

\item{opts}{\code{vector} inputs/reactives referenced in options accompanying expression}
}
\description{
list all reactive invalidators in code block, depending on type and where
  within the code the inputs/reactives are used
}
