% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_contains_app}
\alias{dir_contains_app}
\title{check if directory has shiny files: app.R or server.R}
\usage{
dir_contains_app(dir)
}
\arguments{
\item{dir}{directory to look for shiny app}
}
\description{
check if directory has shiny files: app.R or server.R
}
