% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_inputs}
\alias{extract_inputs}
\title{extract all inputs from shiny server code}
\usage{
extract_inputs(code, input = "input",
  regex = regex_pattern()$input_pattern)
}
\arguments{
\item{code}{\code{string} server.R code}

\item{input}{\code{string} name of reactie input list, defaults to "input"}

\item{regex}{\code{string} regex pattern to detect reactive inputs}
}
\description{
extract all inputs using regex from shiny server code
}
