% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_inputs_with_id}
\alias{extract_inputs_with_id}
\title{extract all inputs from server alongside ids}
\usage{
extract_inputs_with_id(server)
}
\arguments{
\item{server}{\code{data.frame} of all code blocks in server.R code}
}
\description{
get all inputs with assigned ids from parsed server code
}
