% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_isolates}
\alias{extract_isolates}
\title{extract from within code block expressions}
\usage{
extract_isolates(code, type, regex = regex_pattern()$isolate)
}
\arguments{
\item{code}{\code{string} expression of code block}

\item{type}{\code{string} type of code block (observe, reactive, observeEvent, ...)}

\item{regex}{\code{string} regex pattern to detect isolation blocks}
}
\description{
extract and concat isolate bodies from code block expression
}
