% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_reactives}
\alias{extract_reactives}
\title{extract all reactives from shiny server code}
\usage{
extract_reactives(code, blocks, regex = regex_pattern()$input_pattern)
}
\arguments{
\item{code}{\code{string} server.R code}

\item{blocks}{\code{data.frame} of all code blocks in server.R code}

\item{regex}{\code{string} regex pattern to detect reactive inputs}
}
\description{
extract all reactives using regex from shiny server code
}
