% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_sources_with_id}
\alias{extract_sources_with_id}
\title{extract all reactive srouces from server alongside ids}
\usage{
extract_sources_with_id(app)
}
\arguments{
\item{app}{\code{data.frame} of all code blocks, inputs and reatives with ids in server.R code}
}
\description{
get all reactive sources with assigned ids from parsed server code
}
