% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{list_matches}
\alias{list_matches}
\title{list all occuring matches in code block}
\usage{
list_matches(code, matches)
}
\arguments{
\item{code}{\code{string} code of code block}

\item{matches}{\code{vector} of all inputs/reactives}
}
\description{
list all occuring matches in code block
}
