% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{list_side_effects}
\alias{list_side_effects}
\title{list all occuring side effects in code block}
\usage{
list_side_effects(code, matches, reactive_value = "values",
  regex_assgn = regex_pattern()$assignment,
  regex_braces = regex_pattern()$braces_start)
}
\arguments{
\item{code}{\code{string} code of code block}

\item{matches}{\code{vector} of all reactives, both reactiveVal() and values$reactiveValues}

\item{reactive_value}{\code{string} name of reactiveValues() object, defautls to "values"}

\item{regex_assgn}{\code{string} regex to detect assignments}

\item{regex_braces}{\code{string} regex to detect assignment in braces}
}
\description{
list all occuring matches in code block
}
