% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_df}
\alias{list_to_df}
\title{convert named list to data.frame}
\usage{
list_to_df(listfordf, name = "name", value = "value")
}
\arguments{
\item{listfordf}{names list to be converted to data.frame}

\item{name}{column name of names}

\item{value}{column name of values}
}
\description{
convert named list to data.frame
}
