% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mermaid_colors}
\alias{mermaid_colors}
\title{colors for mermaid graph nodes}
\usage{
mermaid_colors(name = "name", value = "color")
}
\arguments{
\item{name}{column name of names}

\item{value}{column name of colors}
}
\description{
colors for mermaid graph nodes
}
