% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_code_blocks}
\alias{parse_code_blocks}
\title{extract code blocks from server.R code}
\usage{
parse_code_blocks(code,
  regex_function = regex_pattern()$function_pattern,
  regex_block = regex_pattern()$shiny_server_blocks)
}
\arguments{
\item{code}{\code{string} server.R code}

\item{regex_function}{\code{string} regex pattern to detect function blocks}

\item{regex_block}{\code{string} regex pattern to detect code blocks}
}
\description{
identify, annotate and extract code blocks in shiny server.R code.
  Annotating the name, function, body, options, expr and isolates
  \code{<name> <- <function>\( <body> <options> \{ <expr> <isolates>\} <options>\) }
}
