% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_shiny_app}
\alias{parse_shiny_app}
\title{parse shiny app located within directory}
\usage{
parse_shiny_app(dir)
}
\arguments{
\item{dir}{\code{string} directory where at least a server.R file is present}
}
\description{
parse shiny app located within directory
}
