% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_shiny_server}
\alias{parse_shiny_server}
\title{parse code of server.R}
\usage{
parse_shiny_server(server_file, dir)
}
\arguments{
\item{server_file}{\code{string} content of server.R file}

\item{dir}{\code{string} directory of server.R file}
}
\description{
parse sever.R code, inline replace source statements, extract
  name of input, output & server; list all used inputs and reactives;
  construct data.frame of seperate code blocks
}
