% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_shiny_server_code}
\alias{parse_shiny_server_code}
\title{extract code within the server definition}
\usage{
parse_shiny_server_code(code, regex = regex_pattern()$shiny_server)
}
\arguments{
\item{code}{\code{string} server.R code}

\item{regex}{\code{string} regex pattern to detect \code{function(<input> , <output>, <sesion> ) { <code> }}}
}
\description{
get shiny code with server defintion \code{function(<input> , <output>, <sesion> ) { <code> }}
  and remove comments
}
