% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_source_statements}
\alias{parse_source_statements}
\title{recursively inline replace shiny source statements}
\usage{
parse_source_statements(code, dir,
  regex = regex_pattern()$source_pattern)
}
\arguments{
\item{code}{\code{string} server.R code}

\item{dir}{\code{string} path of server.R file}

\item{regex}{\code{string} regex pattern to detect source statements}
}
\description{
detect all source statements used to split shiy server into separate
  files and replace them inline with the actual sourced code. Do this recursively
  for sourced files so we make sure there are no source statements left in the
  returned code
}
