% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid.R
\name{plot_mermaid}
\alias{plot_mermaid}
\title{plot mermaid diagram from analyzed shiny app code}
\usage{
plot_mermaid(app, network, ...)
}
\arguments{
\item{app}{\code{data.frame} containing all code blocks, inputs and reactives}

\item{network}{\code{data.frame} containing all reactive dependencies}

\item{...}{optional arguments for \code{\link{render_mermaid}}}
}
\description{
plot mermaid diagram from analyzed shiny app code
}
