% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush.R
\name{reactflush}
\alias{reactflush}
\title{anlalyze your shiny app code and display reactivity network}
\usage{
reactflush(dir = getwd(), ...)
}
\arguments{
\item{dir}{root directory of shiny app}

\item{...}{optional arguments for \code{\link{render_mermaid}}}
}
\description{
anlalyze your shiny app code and display reactivity network
}
\examples{
reactflush(fs::path(system.file("examples", package="shiny"), "01_hello"))
reactflush(fs::path(system.file("examples", package="shiny"), "02_text"))
reactflush(fs::path(system.file("examples", package="shiny"), "03_reactivity"))
reactflush(fs::path(system.file("examples", package="shiny"), "04_mpg"))
reactflush(fs::path(system.file("examples", package="shiny"), "05_sliders"))
reactflush(fs::path(system.file("examples", package="shiny"), "06_tabsets"))
reactflush(fs::path(system.file("examples", package="shiny"), "07_widgets"))
reactflush(fs::path(system.file("examples", package="shiny"), "08_html"))
reactflush(fs::path(system.file("examples", package="shiny"), "09_upload"))
reactflush(fs::path(system.file("examples", package="shiny"), "10_download"))
reactflush(fs::path(system.file("examples", package="shiny"), "11_timer"))

}
