% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush.R
\name{reactflush_focus_addin}
\alias{reactflush_focus_addin}
\title{RStudio addin reactflush with selction focus wrapper function}
\usage{
reactflush_focus_addin()
}
\description{
RStudio addin reactflush with selction focus wrapper function
}
