% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush.R
\name{reactflush_markdown}
\alias{reactflush_markdown}
\title{anlalyze your shiny app code and output reactivity network to markdown}
\usage{
reactflush_markdown(dir = getwd())
}
\arguments{
\item{dir}{root directory of shiny app}
}
\description{
anlalyze your shiny app code and output reactivity network to markdown
}
