% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_recursion}
\alias{regex_recursion}
\title{create recursive regex}
\usage{
regex_recursion(.open, .close, .group = 1, .optional = FALSE,
  .sub = "")
}
\arguments{
\item{.open}{symbol at start of recursion}

\item{.close}{symbol at end of recursion}

\item{.group}{reference group within regex}

\item{.optional}{if match is optional}

\item{.sub}{optional regex sub pattern to match}
}
\description{
create recursive regex
}
