% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mermaid.R
\name{render_mermaid}
\alias{render_mermaid}
\title{generate mermaid code from app code analysis}
\usage{
render_mermaid(app, network, focus, focus_field = c("name", "id",
  "type"))
}
\arguments{
\item{app}{\code{data.frame} containing all code blocks, inputs and reactives}

\item{network}{\code{data.frame} containing all reactive dependencies}

\item{focus}{\code{string} reactive node to focus on, all it's dependencies on blocks which
depend on it will be shown}

\item{focus_field}{\code{enum} what the type of the focus argument is:
\itemize{
  \item \code{name}: assigned name of reactive block (\code{(output$)<name> <- reactive/render*})
  \item \code{id}: assigned id by \code{\link{parse_code_blocks}}
  \item \code{type}: type of block (reactive, observeEvent, ...)
}}
}
\description{
generate mermaid code from app code analysis
}
