% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_match_all_perl}
\alias{str_match_all_perl}
\title{stringr match_all adaptation to work with PCRE engine}
\usage{
str_match_all_perl(s, p, as_tibble = TRUE)
}
\arguments{
\item{s}{string}

\item{p}{pattern}

\item{as_tibble}{return matches as tibble}
}
\description{
Helper function to replicate str_match_all functionality with
PERL like syntax
}
